frontend(onboarding): combined signup and onboarding and adjusted css to fit
