--- frontend/src/app/signup/SignUp.module.css
@@ -1,26 +1,0 @@
-.signup {
-  justify-content: center;
-  align-items: center;
-  display: flex;
-  background-color: white;
-  border-radius: 10px;
-  height: 100%;
-}
-
-.signupTitle {
-  width: 323px;
-  height: 42px;
-}
-
-.signupWelcome {
-  margin-top: -20px;
-  width: 323px;
-  height: 27px;
-}
-
-.signupNames {
-  display: flex;
-  gap: 2rem;
-  margin-top: 20px;
-  width: 100%;
-}
