+++ frontend/src/components/Button.module.css
@@ -18,6 +18,7 @@
 
 .primary:disabled {
   background: var(--color-grey2);
+  pointer-events: none;
 }
 
 .primary:active {
@@ -35,6 +36,7 @@
 
 .secondary:disabled {
   background: var(--color-grey2);
+  pointer-events: none;
 }
 
 .secondary:active {
