+++ frontend/src/components/onboardingForm/Basics.module.css
@@ -1,9 +1,9 @@
 .container {
-  position: fixed;
+  position: absolute;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  width: 60%;
+  width: 75%;
   max-height: 80vh; /* Limit height to 80% of viewport height */
   display: flex;
   flex-direction: column;
@@ -14,7 +14,7 @@
   overflow-y: auto; /* Enable vertical scrolling */
   scrollbar-width: none; /* Firefox */
   -ms-overflow-style: none; /* IE and Edge */
-  margin-left: 9rem;
+  max-width: 750px;
 }
 
 /* Hide scrollbar for Chrome, Safari and Opera */
