+++ frontend/src/components/onboardingForm/Category.module.css
@@ -9,8 +9,8 @@
 
 .container {
   background: var(--color-surface);
-  width: 80%;
-  max-width: 700px;
+  width: 75%;
+  max-width: 750px;
   border-radius: 1rem;
   display: flex;
   flex-direction: column;
