+++ frontend/src/components/onboardingForm/Questionnaire.module.css
@@ -1,20 +1,20 @@
 .container {
-  position: fixed;
+  position: absolute;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  width: 60%;
-  max-height: 80vh;
+  width: 75%;
+  max-height: 80vh; /* Limit height to 80% of viewport height */
   display: flex;
   flex-direction: column;
   align-items: flex-start;
   border-radius: 10px;
   background: var(--color-surface);
   padding: 20px;
-  margin-left: 9rem;
   overflow-y: auto; /* Enable vertical scrolling */
-  scrollbar-width: none;
-  -ms-overflow-style: none;
+  scrollbar-width: none; /* Firefox */
+  -ms-overflow-style: none; /* IE and Edge */
+  max-width: 750px;
 }
 
 .container::-webkit-scrollbar {
