+++ frontend/src/components/onboardingForm/SignUp.module.css
@@ -1,0 +1,63 @@
+.signup {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.signupContainer {
+  width: 75%;
+  max-width: 750px;
+  padding: 20px;
+}
+
+.signupTitle {
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin-bottom: 0.5rem;
+  color: var(--color-primary2);
+}
+
+.signupWelcome {
+  margin-bottom: 2rem;
+}
+
+.signupNames {
+  display: flex;
+  column-gap: 4%;
+  flex-wrap: wrap;
+}
+
+.inputGroup {
+  flex: 1 1 48%;
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 1rem;
+}
+
+.label {
+  margin-bottom: 0.25rem;
+  font-weight: 500;
+}
+
+.input {
+  padding: 0.5rem;
+  border: 1px solid var(--color-grey3);
+  border-radius: 8px;
+  font-size: 1rem;
+}
+
+.errorText {
+  font-size: 0.75rem;
+  color: var(--color-error);
+  margin-top: 0.25rem;
+  min-height: 1em;
+}
+
+.buttonContainer {
+  display: flex;
+  justify-content: flex-end;
+  margin-top: 1rem;
+}
