+++ frontend/src/components/onboardingForm/Student.module.css
@@ -6,7 +6,8 @@
 }
 
 .form {
-  width: 70%;
+  width: 75%;
+  max-width: 750px;
   display: flex;
   flex-direction: column;
   align-items: flex-start;
