+++ frontend/src/components/onboardingForm/index.ts
@@ -4,3 +4,4 @@ export { Questionnaire } from "./Questionnaire";
 export { Student } from "./Student";
 export { Associate } from "./Associate";
 export { Category } from "./Category";
+export { SignUp } from "./Signup";
