+++ frontend/src/state/store.ts
@@ -2,6 +2,12 @@ import { GlobalState, createStore } from "little-state-machine";
 
 const initialState: GlobalState = {
   onboardingForm: {
+    // Signup
+    firstName: "",
+    lastName: "",
+    email: "",
+    password: "",
+    // Basic
     professionalTitle: { value: "", label: "" },
     country: { value: "", label: "" }, // how to make display default
     languages: [],
